+++ packages/blog/gatsby-blog-core/src/fragments/article.information.js
@@ -2,7 +2,20 @@ import { graphql } from 'gatsby'
 
 export const query = graphql`
   fragment ArticleInformation on Article {
-    ...ArticlePreview
+    id
+    title
+    slug
+    link
+    timeToRead @include(if: $includeTimeToRead)
+    featured
+    thumbnailText
+    date(formatString: "MMMM DD, YYYY")
+    category {
+      ...ArticleCategory
+    }
+    author {
+      ...ArticleAuthor
+    }
     body
     keywords
     tags {