+++ packages/blog/gatsby-blog-core/src/templates/collection.category.js
@@ -4,7 +4,14 @@ import Collection from '../containers/Collection'
 export default Collection
 
 export const pageQuery = graphql`
-  query allArticleByCategoryQuery($skip: Int!, $limit: Int!, $slug: String!) {
+  query allArticleByCategoryQuery(
+    $skip: Int!
+    $limit: Int!
+    $slug: String!
+    $includeExcerpt: Boolean!
+    $includeTimeToRead: Boolean!
+    $imageQuality: Int!
+  ) {
     collectionInfo: articleCategory(slug: { eq: $slug }) {
       id
       name
@@ -15,7 +22,7 @@ export const pageQuery = graphql`
     posts: allArticle(
       filter: {
         private: { ne: true }
-        draft: {ne: true}
+        draft: { ne: true }
         category: { slug: { eq: $slug } }
       }
       sort: { fields: [date], order: DESC }