+++ packages/blog/gatsby-blog-core/src/templates/collection.tag.js
@@ -4,7 +4,14 @@ import Collection from '../containers/Collection'
 export default Collection
 
 export const pageQuery = graphql`
-  query allArticleByTagQuery($skip: Int!, $limit: Int!, $slug: String!) {
+  query allArticleByTagQuery(
+    $skip: Int!
+    $limit: Int!
+    $slug: String!
+    $includeExcerpt: Boolean!
+    $includeTimeToRead: Boolean!
+    $imageQuality: Int!
+  ) {
     collectionInfo: articleTag(slug: { eq: $slug }) {
       id
       name
@@ -14,7 +21,7 @@ export const pageQuery = graphql`
     posts: allArticle(
       filter: {
         private: { ne: true }
-        draft: {ne: true}
+        draft: { ne: true }
         tags: { elemMatch: { slug: { eq: $slug } } }
       }
       sort: { fields: [date], order: DESC }