+++ packages/blog/gatsby-blog-core/src/utils/default.options.js
@@ -111,13 +111,20 @@ module.exports = pluginOptions => {
 
   const slugSanitizeRegex = getValue(pluginOptions, 'slugSanitizeRegex', null)
 
+  const includeExcerpt = getValue(pluginOptions, 'includeExcerpt', true)
+  const includeTimeToRead = getValue(pluginOptions, 'includeTimeToRead', true)
+  const imageQuality = getValue(pluginOptions, 'imageQuality', 75)
+
   const pageContextOptions = {
     paginatePostsPage,
     basePath,
     services,
     siteUrl,
     mobileMenu,
-    darkMode
+    darkMode,
+    includeExcerpt,
+    includeTimeToRead,
+    imageQuality
   }
 
   return {