+++ packages/blog/gatsby-blog-pages/src/containers/Collection.Author.jsx
@@ -12,7 +12,7 @@ const PageCollectionAuthors = ({
   ...props
 }) => (
   <Layout {...props}>
-    <Seo title={collectionInfo.name} />
+    <Seo title={collectionInfo.name} description={collectionInfo.description} />
     <Divider />
     <Stack effectProps={{ effect: 'fadeInDown' }}>
       <PageTitle header='Published by Author' totalCount={posts.totalCount} />